% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\docType{class}
\name{EnhancerRNAGene-class}
\alias{EnhancerRNAGene-class}
\title{Class "EnhancerRNAGene"}
\description{
Class "EnhancerRNAGene"
}
\section{Slots}{

\describe{
\item{\code{targetGeneID}}{The unique identifier for the gene targeted by the enhancer RNA (eRNA) (character).}

\item{\code{rnaSequence}}{The nucleotide sequence of the RNA encoded by the enhancer RNA gene (character).}
}}

\seealso{
Other Gene classes: 
\code{\link{LongNonCodingRNAGene-class}},
\code{\link{MicroRNAGene-class}},
\code{\link{ProteinCodingGene-class}},
\code{\link{Pseudogene-class}},
\code{\link{RibosomalRNAGene-class}},
\code{\link{SmallNuclearRNAGene-class}},
\code{\link{TransferRNAGene-class}}
}
\concept{Gene classes}
