% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\name{EnhancerRNAGene}
\alias{EnhancerRNAGene}
\title{Create an EnhancerRNAGene Object}
\usage{
EnhancerRNAGene(
  id,
  hugoSymbol,
  name,
  targetGeneID,
  rnaSequence,
  description = NA_character_,
  structure = GRanges()
)
}
\arguments{
\item{id}{A unique identifier for the gene, such as Ensembl or NCBI gene ID (character).}

\item{hugoSymbol}{The HUGO (Human Genome Organisation) symbol for the gene (character).}

\item{name}{The full name of the gene (character).}

\item{targetGeneID}{The unique identifier for the gene targeted by the enhancer RNA (eRNA) (character).}

\item{rnaSequence}{The nucleotide sequence of the RNA encoded by the eRNA gene (character).}

\item{description}{A brief description of the gene (character). Defaults to \code{NA}.}

\item{structure}{The genomic structure of the gene, typically represented as a \code{GRanges} object indicating chromosome, start, end, and strand. Defaults to an empty \code{GRanges} object.}
}
\value{
An object of class \code{EnhancerRNAGene}.
}
\description{
This function creates an object of class \code{EnhancerRNAGene}, representing an enhancer RNA (eRNA) gene with its attributes.
}
