% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\docType{class}
\name{Gene-class}
\alias{Gene-class}
\title{Class Gene: a class to represent a generic gene with its attributes.}
\description{
Class Gene: a class to represent a generic gene with its attributes.
}
\section{Slots}{

\describe{
\item{\code{id}}{Gene ID (character)}

\item{\code{hugoSymbol}}{HUGO Symbol (character)}

\item{\code{name}}{Gene name (character)}

\item{\code{description}}{Description of the gene (character)}

\item{\code{structure}}{Gene structure, typically as a \code{GRanges} object}
}}

