% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\name{Gene}
\alias{Gene}
\title{Create a Gene Object}
\usage{
Gene(id, symbol, name, description, structure, product)
}
\arguments{
\item{id}{A unique identifier for the gene, such as Ensembl or NCBI gene ID (character).}

\item{symbol}{The HUGO (Human Genome Organisation) symbol for the gene (character).}

\item{name}{The full name of the gene (character).}

\item{description}{A brief description of the gene (character).}

\item{structure}{The genomic structure of the gene, typically represented as a \code{GRanges} object indicating chromosome, start, end, and strand.}

\item{product}{A list containing gene products, such as protein sequences or RNA sequences.}
}
\value{
An object of class \code{Gene}.
}
\description{
This function creates an object of class \code{Gene}, representing a generic gene with its attributes.
}
