% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\docType{methods}
\name{hugoSymbol}
\alias{hugoSymbol}
\alias{hugoSymbol,Gene-method}
\alias{hugoSymbol<-}
\alias{hugoSymbol<-,Gene-method}
\title{Get the HUGO Symbol of a Gene Object
This function retrieves the HUGO Symbol from an object of class \code{Gene}.}
\usage{
hugoSymbol(gene)

\S4method{hugoSymbol}{Gene}(gene)

hugoSymbol(gene) <- value

\S4method{hugoSymbol}{Gene}(gene) <- value
}
\arguments{
\item{gene}{An object of class \code{Gene}.}

\item{value}{The HUGO Symbol to assign to the gene (character).}
}
\value{
The HUGO Symbol associated with the gene (character).

The updated gene object with the new HUGO Symbol.
}
\description{
Get the HUGO Symbol of a Gene Object
This function retrieves the HUGO Symbol from an object of class \code{Gene}.

Set the HUGO Symbol of a Gene Object
This function sets the HUGO Symbol for an object of class \code{Gene}.
}
