% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\docType{class}
\name{LongNonCodingRNAGene-class}
\alias{LongNonCodingRNAGene-class}
\title{Class "LongNonCodingRNAGene"}
\description{
Class "LongNonCodingRNAGene"
}
\section{Slots}{

\describe{
\item{\code{lncRNAID}}{A unique identifier for the long non-coding RNA (lncRNA) product (character).}

\item{\code{rnaSequence}}{The nucleotide sequence of the RNA encoded by the lncRNA gene (character).}
}}

\seealso{
Other Gene classes: 
\code{\link{EnhancerRNAGene-class}},
\code{\link{MicroRNAGene-class}},
\code{\link{ProteinCodingGene-class}},
\code{\link{Pseudogene-class}},
\code{\link{RibosomalRNAGene-class}},
\code{\link{SmallNuclearRNAGene-class}},
\code{\link{TransferRNAGene-class}}
}
\concept{Gene classes}
