% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\name{LongNonCodingRNAGene}
\alias{LongNonCodingRNAGene}
\title{Create a LongNonCodingRNAGene Object}
\usage{
LongNonCodingRNAGene(
  id,
  hugoSymbol,
  name,
  lncRNAID,
  rnaSequence,
  description = NA_character_,
  structure = GRanges()
)
}
\arguments{
\item{id}{A unique identifier for the gene, such as Ensembl or NCBI gene ID (character).}

\item{hugoSymbol}{The HUGO (Human Genome Organisation) symbol for the gene (character).}

\item{name}{The full name of the gene (character).}

\item{lncRNAID}{A unique identifier for the long non-coding RNA (lncRNA) product (character).}

\item{rnaSequence}{The nucleotide sequence of the RNA encoded by the lncRNA gene (character).}

\item{description}{A brief description of the gene (character). Defaults to NA.}

\item{structure}{The genomic structure of the gene, typically represented as a \code{GRanges} object. Defaults to an empty \code{GRanges} object.}
}
\value{
An object of class \code{LongNonCodingRNAGene}.
}
\description{
This function creates an object of class \code{LongNonCodingRNAGene}, representing a long non-coding RNA gene with additional attributes.
}
