% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\docType{class}
\name{MicroRNAGene-class}
\alias{MicroRNAGene-class}
\title{Class "MicroRNAGene"}
\description{
Class "MicroRNAGene"
}
\section{Slots}{

\describe{
\item{\code{miRNAID}}{A unique identifier for the microRNA (miRNA) product (character).}

\item{\code{rnaSequence}}{The nucleotide sequence of the RNA encoded by the miRNA gene (character).}

\item{\code{seedSequence}}{The seed sequence of the miRNA, typically a crucial part of miRNA-target recognition (character).}
}}

\seealso{
Other Gene classes: 
\code{\link{EnhancerRNAGene-class}},
\code{\link{LongNonCodingRNAGene-class}},
\code{\link{ProteinCodingGene-class}},
\code{\link{Pseudogene-class}},
\code{\link{RibosomalRNAGene-class}},
\code{\link{SmallNuclearRNAGene-class}},
\code{\link{TransferRNAGene-class}}
}
\concept{Gene classes}
