% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\name{MicroRNAGene}
\alias{MicroRNAGene}
\title{Create a MicroRNAGene Object}
\usage{
MicroRNAGene(
  id,
  hugoSymbol,
  name,
  miRNAID,
  seedSequence,
  description = NA_character_,
  structure = GRanges()
)
}
\arguments{
\item{id}{A unique identifier for the gene, such as Ensembl or NCBI gene ID (character).}

\item{hugoSymbol}{The HUGO (Human Genome Organisation) symbol for the gene (character).}

\item{name}{The full name of the gene (character).}

\item{miRNAID}{A unique identifier for the microRNA (miRNA) product (character).}

\item{seedSequence}{The seed sequence of the miRNA, typically a crucial part of miRNA-target recognition (character).}

\item{description}{A brief description of the gene (character). Defaults to NA.}

\item{structure}{The genomic structure of the gene, typically represented as a \code{GRanges} object. Defaults to an empty \code{GRanges} object.}
}
\value{
An object of class \code{MicroRNAGene}.
}
\description{
This function creates an object of class \code{MicroRNAGene}, representing a microRNA gene with additional attributes.
}
