% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\docType{class}
\name{ProteinCodingGene-class}
\alias{ProteinCodingGene-class}
\title{Class "ProteinCodingGene"}
\description{
Class "ProteinCodingGene"
}
\section{Slots}{

\describe{
\item{\code{proteinID}}{A unique identifier for the protein product encoded by the gene (character).}

\item{\code{proteinSequence}}{The amino acid sequence of the protein encoded by the gene (character).}

\item{\code{exonCount}}{The number of exons in the gene (integer).}
}}

\seealso{
Other Gene classes: 
\code{\link{EnhancerRNAGene-class}},
\code{\link{LongNonCodingRNAGene-class}},
\code{\link{MicroRNAGene-class}},
\code{\link{Pseudogene-class}},
\code{\link{RibosomalRNAGene-class}},
\code{\link{SmallNuclearRNAGene-class}},
\code{\link{TransferRNAGene-class}}
}
\concept{Gene classes}
