% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\name{ProteinCodingGene}
\alias{ProteinCodingGene}
\title{Create a ProteinCodingGene Object}
\usage{
ProteinCodingGene(
  id,
  hugoSymbol,
  name,
  proteinID,
  proteinSequence,
  exonCount,
  description = NA_character_,
  structure = GRanges()
)
}
\arguments{
\item{id}{A unique identifier for the gene, such as Ensembl or NCBI gene ID (character).}

\item{hugoSymbol}{The HUGO (Human Genome Organisation) symbol for the gene (character).}

\item{name}{The full name of the gene (character).}

\item{proteinID}{A unique identifier for the protein product encoded by the gene (character).}

\item{proteinSequence}{The amino acid sequence of the protein encoded by the gene (character).}

\item{exonCount}{The number of exons in the gene (integer).}

\item{description}{A brief description of the gene (character). Defaults to NA.}

\item{structure}{The genomic structure of the gene, typically represented as a \code{GRanges} object. Defaults to an empty \code{GRanges} object.}
}
\value{
An object of class \code{ProteinCodingGene}.
}
\description{
This function creates an object of class \code{ProteinCodingGene}, representing a protein-coding gene with additional attributes.
}
