% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\docType{class}
\name{Pseudogene-class}
\alias{Pseudogene-class}
\title{Class "Pseudogene"}
\description{
Class "Pseudogene"
}
\section{Slots}{

\describe{
\item{\code{parentGeneID}}{The unique identifier for the parent gene from which the pseudogene originated (character).}

\item{\code{functionalStatus}}{The functional status of the pseudogene, indicating whether it is transcribed, translated, or non-functional (character).}
}}

\seealso{
Other Gene classes: 
\code{\link{EnhancerRNAGene-class}},
\code{\link{LongNonCodingRNAGene-class}},
\code{\link{MicroRNAGene-class}},
\code{\link{ProteinCodingGene-class}},
\code{\link{RibosomalRNAGene-class}},
\code{\link{SmallNuclearRNAGene-class}},
\code{\link{TransferRNAGene-class}}
}
\concept{Gene classes}
