% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\name{Pseudogene}
\alias{Pseudogene}
\title{Create a Pseudogene Object}
\usage{
Pseudogene(
  id,
  hugoSymbol,
  name,
  parentGeneID,
  functionalStatus,
  description = NA_character_,
  structure = GRanges()
)
}
\arguments{
\item{id}{A unique identifier for the gene, such as Ensembl or NCBI gene ID (character).}

\item{hugoSymbol}{The HUGO (Human Genome Organisation) symbol for the gene (character).}

\item{name}{The full name of the gene (character).}

\item{parentGeneID}{The unique identifier for the parent gene from which the pseudogene originated (character).}

\item{functionalStatus}{The functional status of the pseudogene, indicating whether it is transcribed, translated, or non-functional (character).}

\item{description}{A brief description of the gene (character). Defaults to \code{NA}.}

\item{structure}{The genomic structure of the gene, typically represented as a \code{GRanges} object indicating chromosome, start, end, and strand. Defaults to an empty \code{GRanges} object.}
}
\value{
An object of class \code{Pseudogene}.
}
\description{
This function creates an object of class \code{Pseudogene}, representing a pseudogene with its attributes.
}
