% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\docType{class}
\name{RibosomalRNAGene-class}
\alias{RibosomalRNAGene-class}
\title{Class "RibosomalRNAGene"}
\description{
Class "RibosomalRNAGene"
}
\section{Slots}{

\describe{
\item{\code{rRNAID}}{A unique identifier for the ribosomal RNA (rRNA) product (character).}

\item{\code{rnaSequence}}{The nucleotide sequence of the RNA encoded by the rRNA gene (character).}
}}

\seealso{
Other Gene classes: 
\code{\link{EnhancerRNAGene-class}},
\code{\link{LongNonCodingRNAGene-class}},
\code{\link{MicroRNAGene-class}},
\code{\link{ProteinCodingGene-class}},
\code{\link{Pseudogene-class}},
\code{\link{SmallNuclearRNAGene-class}},
\code{\link{TransferRNAGene-class}}
}
\concept{Gene classes}
