% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\docType{class}
\name{SmallNuclearRNAGene-class}
\alias{SmallNuclearRNAGene-class}
\title{Class "SmallNuclearRNAGene"}
\description{
Class "SmallNuclearRNAGene"
}
\section{Slots}{

\describe{
\item{\code{snRNAID}}{A unique identifier for the small nuclear RNA (snRNA) product (character).}

\item{\code{rnaSequence}}{The nucleotide sequence of the RNA encoded by the snRNA gene (character).}
}}

\seealso{
Other Gene classes: 
\code{\link{EnhancerRNAGene-class}},
\code{\link{LongNonCodingRNAGene-class}},
\code{\link{MicroRNAGene-class}},
\code{\link{ProteinCodingGene-class}},
\code{\link{Pseudogene-class}},
\code{\link{RibosomalRNAGene-class}},
\code{\link{TransferRNAGene-class}}
}
\concept{Gene classes}
