% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\name{SmallNuclearRNAGene}
\alias{SmallNuclearRNAGene}
\title{Create a SmallNuclearRNAGene Object}
\usage{
SmallNuclearRNAGene(
  id,
  hugoSymbol,
  name,
  snRNAID,
  rnaSequence,
  description = NA_character_,
  structure = GRanges()
)
}
\arguments{
\item{id}{A unique identifier for the gene, such as Ensembl or NCBI gene ID (character).}

\item{hugoSymbol}{The HUGO (Human Genome Organisation) symbol for the gene (character).}

\item{name}{The full name of the gene (character).}

\item{snRNAID}{A unique identifier for the small nuclear RNA (snRNA) product (character).}

\item{rnaSequence}{The nucleotide sequence of the RNA encoded by the snRNA gene (character).}

\item{description}{A brief description of the gene (character). Defaults to NA.}

\item{structure}{The genomic structure of the gene, typically represented as a \code{GRanges} object. Defaults to an empty \code{GRanges} object.}
}
\value{
An object of class \code{SmallNuclearRNAGene}.
}
\description{
This function creates an object of class \code{SmallNuclearRNAGene}, representing a small nuclear RNA gene with additional attributes.
}
