% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\docType{class}
\name{TransferRNAGene-class}
\alias{TransferRNAGene-class}
\title{Class "TransferRNAGene"}
\description{
Class "TransferRNAGene"
}
\section{Slots}{

\describe{
\item{\code{tRNAID}}{A unique identifier for the transfer RNA (tRNA) product (character).}

\item{\code{aminoAcid}}{The amino acid carried by the tRNA (character).}

\item{\code{rnaSequence}}{The nucleotide sequence of the RNA encoded by the tRNA gene (character).}
}}

\seealso{
Other Gene classes: 
\code{\link{EnhancerRNAGene-class}},
\code{\link{LongNonCodingRNAGene-class}},
\code{\link{MicroRNAGene-class}},
\code{\link{ProteinCodingGene-class}},
\code{\link{Pseudogene-class}},
\code{\link{RibosomalRNAGene-class}},
\code{\link{SmallNuclearRNAGene-class}}
}
\concept{Gene classes}
