% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\name{TransferRNAGene}
\alias{TransferRNAGene}
\title{Create a TransferRNAGene Object}
\usage{
TransferRNAGene(
  id,
  hugoSymbol,
  name,
  tRNAID,
  aminoAcid,
  rnaSequence,
  description = NA_character_,
  structure = GRanges()
)
}
\arguments{
\item{id}{A unique identifier for the gene, such as Ensembl or NCBI gene ID (character).}

\item{hugoSymbol}{The HUGO (Human Genome Organisation) symbol for the gene (character).}

\item{name}{The full name of the gene (character).}

\item{tRNAID}{A unique identifier for the transfer RNA (tRNA) product (character).}

\item{aminoAcid}{The amino acid carried by the tRNA (character).}

\item{rnaSequence}{The nucleotide sequence of the RNA encoded by the tRNA gene (character).}

\item{description}{A brief description of the gene (character). Defaults to \code{NA}.}

\item{structure}{The genomic structure of the gene, typically represented as a \code{GRanges} object indicating chromosome, start, end, and strand. Defaults to an empty \code{GRanges} object.}
}
\value{
An object of class \code{TransferRNAGene}.
}
\description{
This function creates an object of class \code{TransferRNAGene}, representing a transfer RNA (tRNA) gene with its attributes.
}
